created custom classes for building and unit stats,, then grabbed them from list of building/unit prefabs and assigned values on GameManager startup. can now use these values to apply tech upgrades to existing entities, tested buidlings and it worke fine, still need to finish units version and will test it all again

+++ Assets/Scripts/UnitScripts/UnitStateController.cs
@@ -31,8 +31,8 @@ public class UnitStateController : Entities
 	public GameObject FoVMeshObj;
 
 	[Header("Unit Stats")]
-	public float attackRange;
-	public float ViewRange;
+	public int attackRange;
+	public int ViewRange;
 
 	[Header("Unit Bools")]
 	public bool isUnitArmed;
